requirements.txt
# =====================================================
# DEPENDENCIAS - AGENTE LANGGRAPH MULTI-TENANT
# =====================================================

# Core LangGraph e LangChain
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Banco de Dados PostgreSQL
asyncpg>=0.29.0
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0
sqlalchemy>=2.0.0

# FastAPI (API REST)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# HTTP Client
httpx>=0.26.0
tenacity>=8.2.0

# Validacao de Dados
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Autenticacao JWT
pyjwt>=2.8.0
bcrypt>=4.1.0

# Supabase (config tables)
supabase>=2.3.0

# Scheduler (follow-ups)
apscheduler>=3.10.0

# Utilidades
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2024.1

# Logs
structlog>=24.1.0
python-json-logger>=2.0.7
